Rename Spec to MSBuildSpec and make public
Move all optional arguments to the end. Add comments.
